Split client into annotations and predictions flavors
