+++ butler/annotations/__init__.py
@@ -1,4 +1,5 @@
 from butler.annotations.annotations import Annotations
+from butler.annotations.annotations_client import AnnotationsClient
 from butler.annotations.bbx_utils import butler_bbx_to_min_max
 from butler.annotations.document_annotation_with_image import DocumentAnnotationWithImage
 from butler.annotations.layoutlm_utils import normalize_ner_annotation_for_layoutlm
