+++ butler/predictions/__init__.py
@@ -1,0 +1,2 @@
+from butler.generated.models import ExtractionResultsDto
+from butler.predictions.predictions_client import PredictionsClient
