+++ butler/test/test_load_annotations.py
@@ -4,7 +4,7 @@ import os
 from butler import Client
 from butler.annotations.layoutlm_utils import normalize_ner_annotation_for_layoutlm
 
-# Run using 'python -m butler.test_load_annotations'
+# Run using 'python -m butler.test.test_load_annotations'
 
 logging.basicConfig(level=logging.INFO)
